# pairwise2: Allow insertion after deletion and performance improvements