Updated <code>pairwise2</code> module:

1. Allow insertion after deletion (and vice versa). An alignment of this type:
      ```

A-
-B
```
is now possible (depending on the gap open penalties), like in EMBOSS' needle. This change was inspired by this [pre-print paper](http://biorxiv.org/content/early/2015/11/12/031500) and this [list](http://www.exelixis-lab.org/web/software/alignment/).

2. Improvements in speed and maximum sequence length for both the pure Python implementation and the C module. (#297, #603, #631). Sequences may be at least 3 times longer before raising a  memory error (on 32 bit Python, which has a maximum memory of 2 GByte), and speed can be up to 10 times faster (see table below).

3. Solved problems with wrong local alignments  (#766, #279).

4. Raise warning, when not using the C module ( #657).

Comparison of old and new <code>pairwise2</code> on 32 bit Python 2.7.11 (with 2 GByte memory limit), Win7 64 bit, Intel Core i5, 2.8 Ghz

 ```
#bad case
length_of_sequence = ...
seq1 = ''.join(['A',]+['I' for x in range(length_of_sequence)])
seq2 = ''.join(['V',]+['I' for x in range(length_of_sequence)])
aln = pairwise2.align.globalxx(seq1, seq2)

 #"real" conditions
seq1 = '...'[:length_of_sequence]  # Just two normal DNA sequences, e.g. human
seq2 = '...'[:length_of_sequence]  # and green monkey dystrophin mRNA (>13 kBp)
aln = pairwise2.align.globalms(seq1, seq2, 5, -4, -10, -1)
 ```

|               | old pairwise2|   --           | new pairwise2|   --     
|---------------|--------------|----------------|--------------|--------
|               | with C module|without C module| with C module| without C module|
|bad case:      |              |                |
|max. length    | ~445         | ~1,050         |   ~7,450     | ~10,200
|time (max len) | 9 sec        | 110 sec        |10 sec        | 730 sec
|time (445 chars)| 9 sec       | 6 sec          | 0.05 sec     | 1.3 sec
|score_only:    |          |||
|max. length    | ~710         | ~1,050         | ~8,430       | ~11,400
|time (max len) | 1.5 sec      | 110 sec        | 6 sec        | 300 sec
|real conditions:||||
|max. length    | ~1,900       | ~3,625         |   ~7,450     | ~10,000
|time (max len) | 10 sec       | 140 sec        | 12 sec       | 690 sec
|time (1900 chars)| 10 sec     | 35 sec         | 1 sec        | 25 sec
|score_only:    |          |||
|max. length    | ~3,230       | ~3,625         | ~ 8,420      | ~11,250
|time (max len) |  4 sec       |  140 sec       | 6 sec        | 300 sec