@MarkusPiotrowski 2016-03-14
@peterjc Sorry for the new style class issue, I was working since November on <code>pairwise2</code> and somehow I missed your commit from December

@mdehoon Let me know, if I can assist you. The three main methods of the module, <code>_make_score_matrix_generic</code>, <code>_make_score_matrix_fast</code> (and it's C implementation) and <code>_recover_alignments</code> were completely rewritten. 
I compared the results for the old and new <code>pairwise2</code> for several thousand random sequences (mostly using <code>align.globalms</code> with different penalties)  and another few examples from reported issues (from here and elsewhere) with positiv results.
1. For global aligments, in most cases the scores of the old and new method stays the same, sometimes it is better with the new method, but it should never be worse.
2. If the score stays the same (in global alignments), the number of alignments is either equal or higher with the new method, it should never be less.
3. This doesn't hold true for local alignments, since there are some issues with the old method (e.g. #766 or here:[http://stackoverflow.com/questions/35874826/biopython-local-alignment-between-dna-sequences-doesnt-find-optimal-alignment?rq=1](url)). Thus the new method may return a worse score or less alignments with the same score, but these should be correct.