@peterjc 2016-03-14
Thanks @MarkusPiotrowski.

(If Michiel doesn't have time to review this at the moment, please remind me we should try to look at this next week)