@mdehoon 2016-03-16
@peterjc , @MarkusPiotrowski  : Thanks for your contribution.
I think that the pairwise2 code has not had a dedicated maintainer since 2005, and while we have done some maintenance on it in the past, I don't think we currently have a real expert for this code among the Biopython developers. So I would suggest to accept the changes by @MarkusPiotrowski  as is, as he has looked at the code in much detail and is likely the most familiar with the code now.

I also think that the user interface to pairwise2 is rather unpythonic, with arguments encoded in the function names, and it currently lacks documentation. I would be in favor of a further revision of this module (after accepting the current changes) to make it more intuitive and pythonic to users. @MarkusPiotrowski  if you'd like to further revise this module, let's discuss on the mailing list.

The only comment on the submitted code is that some functions in cpairwise2module.c (for example calc_affine_penalty) could be declared static if they are not used outside of this file.