@MarkusPiotrowski 2016-03-22
[This question](http://stackoverflow.com/questions/36142371/nucleotides-separator-in-the-pairwise-sequence-alignment-bio-python/36158674#36158674) on Stack Overflow made me clear why it may be a good idea that <code>pairwise2</code> can accept and handle <code>lists</code> as input sequence.

Unfortunately, this does not work any longer (<code>lists</code> are accepted as input but are then converted to strings).. fortunately, it seems that I can repair this.