@peterjc 2016-03-22
If you think it makes sense to deal with lists of elements like that as in my Stack Overflow answer (see also #789), that's fine. Personally I felt I was tricking the code and it was just luck it worked.