@MarkusPiotrowski 2016-03-22
Actually, it was the intended behaviour to accept lists (and list like object), it's written in the comments of the <code>_recover_alignments</code> method. But I didn't understand what this should be good for. The Stack Overflow question opened my eyes, this is exactly the use case (btw, the second answer is from me).
Again, another proof that the modules documentation must be improved. <code>pairwise2</code> has really some nice/powerfull features that EMBOSS needle/water don't have (e.g. to define your own gap function).