@MarkusPiotrowski 2016-04-01
OK, I'm ready. Any chance to get this still into <code>Biopython 1.67</code>? Would be great!