@MarkusPiotrowski 2016-05-21
Merged upstream updates and squashed commits. Should make it easier to merge... ;-)