@MarkusPiotrowski 2016-06-06
The @codecov-io comment above (March 11) is not longer useful when merging recent upstream master into the pull request. It seems to compare the actual (updated) pull request with the upstream master at the time of the first pull request?