@peterjc 2016-06-06
I plan to release Biopython 1.67 in the next couple of days, and then have this PR be the first thing we include in Biopython 1.68.