@peterjc 2016-06-15
Could you touch up the C file to remove trailing white space, and convert mixed indentation to use four spaces (which seems to be the most commonly used style in our other C code)?