@peterjc 2016-06-15
Also is this (Apple) compiler warning easy to address?

```
running build_ext
building 'Bio.cpairwise2' extension
cc -DNDEBUG -g -fwrapv -Os -Wall -Wstrict-prototypes -Qunused-arguments -Qunused-arguments -arch i386 -arch x86_64 -pipe -I/System/Library/Frameworks/Python.framework/Versions/2.6/include/python2.6 -c Bio/cpairwise2module.c -o build/temp.macosx-10.11-intel-2.6/Bio/cpairwise2module.o
Bio/cpairwise2module.c:388:9: warning: add explicit braces to avoid dangling else [-Wdangling-else]
        Py_DECREF(py_score_matrix);
        ^
/System/Library/Frameworks/Python.framework/Versions/2.6/include/python2.6/object.h:762:5: note: expanded from macro 'Py_DECREF'
    else                                                \
    ^
Bio/cpairwise2module.c:390:9: warning: add explicit braces to avoid dangling else [-Wdangling-else]
        Py_DECREF(py_trace_matrix);
        ^
/System/Library/Frameworks/Python.framework/Versions/2.6/include/python2.6/object.h:762:5: note: expanded from macro 'Py_DECREF'
    else                                                \
    ^
2 warnings generated.
Bio/cpairwise2module.c:388:9: warning: add explicit braces to avoid dangling else [-Wdangling-else]
        Py_DECREF(py_score_matrix);
        ^
/System/Library/Frameworks/Python.framework/Versions/2.6/include/python2.6/object.h:762:5: note: expanded from macro 'Py_DECREF'
    else                                                \
    ^
Bio/cpairwise2module.c:390:9: warning: add explicit braces to avoid dangling else [-Wdangling-else]
        Py_DECREF(py_trace_matrix);
        ^
/System/Library/Frameworks/Python.framework/Versions/2.6/include/python2.6/object.h:762:5: note: expanded from macro 'Py_DECREF'
    else                                                \
    ^
2 warnings generated.
cc -Wl,-F. -bundle -undefined dynamic_lookup -arch i386 -arch x86_64 -Qunused-arguments -Qunused-arguments build/temp.macosx-10.11-intel-2.6/Bio/cpairwise2module.o -o build/lib.macosx-10.11-intel-2.6/Bio/cpairwise2.so
```