@MarkusPiotrowski 2016-06-15
Did my best on the C code and obiously I didn't break it... @peterjc: can you check the compiler warnings?

There are some issues that may be closed now: #279, #297, #603, #766 
Some pull requests may be obsolete?  #631, #652