@peterjc 2016-06-15
Cherry-picked as https://github.com/biopython/biopython/commit/29568172fbb381a101ac7efba76bb1cfff67eef2 - no more warnings compiling it on my Mac laptop - good job 👍